backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt
psycopg2-binary>=2.8,<3.0
requests>=2.25,<3.0
django-cors-headers
dj-database-url